refactor(router): misc refactoring
