+++ packages/router/test/apply_redirects.spec.ts
@@ -12,8 +12,7 @@ import {Observable} from 'rxjs/Observable';
 import {of } from 'rxjs/observable/of';
 
 import {applyRedirects} from '../src/apply_redirects';
-import {Routes} from '../src/config';
-import {LoadedRouterConfig} from '../src/router_config_loader';
+import {LoadedRouterConfig, Routes} from '../src/config';
 import {DefaultUrlSerializer, UrlSegmentGroup, UrlTree, equalSegments} from '../src/url_tree';
 
 describe('applyRedirects', () => {
